les modules se sauvegardent et se chargent

+++ synthlabGUI/src/main/java/group1/project/synthlab/ihm/module/IPModule.java
@@ -9,5 +9,5 @@ public interface IPModule extends IPModuleObservable, Serializable {
 	public void stop();
 	public Point getLocation();
 	public void updatePresentation();
-	public void updateLocation(double x, double y);
+	public void updateLocation(int x, int y);
 }
